school_task/requirements.txt
asgiref==3.3.1
automium==0.2.6
automium-web==0.1.1
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
Django==3.1.7
djangorestframework==3.12.2
Flask==1.1.2
gunicorn==20.0.4
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.3
legacy==0.1.6
MarkupSafe==1.1.1
mysqlclient==2.0.3
pytz==2021.1
requests==2.25.1
sqlparse==0.4.1
urllib3==1.26.3
Werkzeug==1.0.1
whitenoise==5.2.0
